+++ packages/antwerp-ui/react-components/src/lib/atoms/radio/Radio.types.ts
@@ -10,6 +10,7 @@ export interface RadioProps {
   onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
   qa?: string;
   value: string;
+  required?: boolean;
 }
 
 export interface RadioGroupProps {
@@ -21,4 +22,5 @@ export interface RadioGroupProps {
   qa?: string;
   state?: 'success' | 'error';
   description?: string;
+  required?: boolean;
 }