+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `required` property to the `Checkbox` component
 - Added `@testing-library/jest-dom` to the dependencies
 - Added `required` prop to the `Checkbox` component
+- Added `required` prop to the `RadioGroup` component
+- Added `required` prop to the `Radio` component (not visible in the UI)
 
 ### Fixed